ai-service/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-jose[cryptography]
python-dotenv